2nd refactor pass.
After we made OfferUtils ortogonal, I've proceeded to formalize the
commonalities between the different Affiliate SDKs and created an
`Affiliate` interface as well as a `InterstitialListener` interface.

I've broken all the static methods that existed in `OfferUtils`, now
called `Offers` and created `Affiliate` implementations for each one
of the sdks.