Removed ugly mime type hack from UIUtils.openFile()
The hack would mess with the given parameters.

The parameters passed should be good from the beginning, so I've added
a method that fixes null mime types for .torrent files (for some reason)
these come with no mime types (maybe there's a further fix at the root
of the issue), and it ensures that `FileDescriptor`'s of `.apk`'s have the right
mime type set before calling `UIUtils.openFile()`.